Add function to get deep immutable version of skylark values
Summary:
For user defined providers, we want to make sure we're only taking immutable versions of skylark values/primitives so they can't change out from under us. As these are only going to be used by user defined rules, we also don't want to have to do a lot of conversion back and forth between skylark types and built in java types (e.g. immutable collections).

So, for the common case where we have a primitive value or an immutable value (e.g. constants defined in the .bzl file, structs, nested sets, etc), just return the original value
For lists, go through and make an immutable list, converting each sub-element (this should normally follow the case outlined above, where it's a list of primitives or immutables)
For dicts, make sure that the keys and values are all immutable, and return an immutable skylarkdict.

For anything else, there isn't a nice way to make it immutable, so fail.

Reviewed By: bobyangyf

shipit-source-id: 4f2edad649