Add a FocusedTargetMatcher that supports build target patterns and regex
Summary:
The previous `FocusedModuleTargetMatcher` had a few issues:
- only supported [build target patterns](https://buck.build/concept/build_target_pattern.html)
- it would find all the targets that matched the pattern and add it to a set, which was used for comparison
   - parsing the graph for all matching targets is unnecessary and slow

Instead, it would be useful to create a real "Matcher", which given a target, can respond with whether it matches or not. This allows us to support regular expressions and test targets on demand vs parsing the whole tree to get targets upfront. By supporting regular expressions, `arc focus` won't need to do a buck query and then find matching focus targets to provide to `buck project`.

The FocusedTargetMatcher parses the space separated focus entry and attempts to parse build target patterns. If it doesn't find a build target pattern, it treats the string as a regular expression.

Currently unused, will be used in next diff.

Reviewed By: williamtwilson

shipit-source-id: aa059b2ffc