Add config option for compiling Kotlin against ABIs
Summary: Compiling against ABIs doesn't currently work as inline methods are not available as they have been stripped away (that will come in future diffs), but it's useful to have this config option for testing with and we'll need it in future anyway. It defaults to false so should have no effect right now.

Reviewed By: asp2insp

shipit-source-id: 0436ba4778