Support for new Action Tags format in requirements files
Summary:
Adding support for more verbose and accurate format used in Worker Requirements files - key is now an ActionTags struct instead of simple rulename. This allows more accurately specifying requirements for Actions.
Also note that if we don't have an entry for an Action including its custom auxiliaryBuildTag, we look for requirements for the rulename without auxiliaryBuildTag (likely we have that entry as they are more common) - this preserves the current behaviour where we ignore the auxiliaryBuildTag entirely.

Reviewed By: alisdair04

shipit-source-id: a77dbc2422