Set SWIFT_INCLUDE_PATHS based on swift_version only
Summary:
We currently rely on the presence of a `.swift` file in the target's inputs to determine whether or not to add the `SWIFT_INCLUDE_PATHS` setting, which is necessary for importing dependencies. However, this breaks on targets that use generated code (via `genrule`) as their only `srcs`, because those are not detected as inputs. Thus, those targets do not compile if they import any non-Apple libraries.

Rather than complicating the logic to detect `genrule`s as well, I think it's safe to just remove this component of the conditional. There are two reasons:

1. We still require a `swift_version` argument, which expresses the intention to use Swift.
2. Even if we didn't, this setting is harmless in Objective-C anyways, it won't change Xcode's behavior, as it only applies to Swift.

This allows targets entirely made up of generated code to compile in Xcode.

shipit-source-id: eba383c98d