Add isKotlinModule to StubJar processing
Summary:
Kotlin ABIs need creating in a different way to normal java ABIs, because Kotlin methods can be `inlined` which means that we need to keep the method body around for the compiler to use when it is doing the inlining.

An easy way to see whether a module contains any Kotlin code is just to look for a `.kotlin_module` file. If this exists, then we want to process the classes a bit differently.

Reviewed By: jkeljo

shipit-source-id: 41d2ab422f