Add copy of Kotlin stdlib
Summary:
When using `kotlinx-metadata-jvm` in a future diff to read the Kotlin metadata off `.class` files, we get the following crash when running: P78695019. The correct solution seems to be to add the kotlin-stdlib, per https://stackoverflow.com/questions/51563273/kotlin-typecastexception-when-trying-to-create-okhttpclient-object. It fixes the issue as expected.

Here I'm just adding the library for future use.

Reviewed By: asp2insp

shipit-source-id: 91f5875d3b