Add option for whether to skip code or not
Summary:
With Kotlin ABIs, methods can be marked as `inline` in which case we need to keep the full method body. In those cases, we don't want to use `SKIP_CODE`, as we always use now.

This diff gives the option to retain the code, but for now that option is not used anywhere. It will be used in future diffs but this diff should not change any behaviour.

Reviewed By: asp2insp

shipit-source-id: b238e4f13c