Fix TargetsCommand with --show-output and android_binary
Summary:
This change modifies `targets` command to always output targets that represent configuration rules with omitting information that is not available for these rules (for example, output file name and rule key).

This change allows to avoid removing configuration nodes from the target graph which fixes that situation when action graph creation stage needs to access configuration nodes (this is happening for `android_binary`).

Reviewed By: mykola-semko

shipit-source-id: b2480b3845