Add Location to DeclaredArtifact
Summary: Optionally pass Location down to artifacts when they are declared, and print that location, if present, in toString(). This is helpful in the case where a user declares a file in a .bzl file and misuses it afterward. We can then say something like "You cannot use the same file, declared at line X, as an output for two different actions".

Reviewed By: bobyangyf

shipit-source-id: ec1b1ff4c1