Move directoriesToMerge logic to a PythonSymlinkTree
Summary: The only place that we use directory merging logic is in PythonInplaceBinary, so move that logic to a special PythonSymlinkTree. This makes non-python usage simpler as well. This also required a lot of changes to un-thread the rule finder that previously was necessary but is now only necessary for python

Reviewed By: cjhopman

shipit-source-id: ed48b4f919