Allow deleting existing symlinks in merge step
Summary: For python binaries, we generate a bunch of empty `__init__.py` files to indicate that these are modules. Unfortunately if a whl has a similar `__init__.py` in the same destination path, we get errors when trying to merge everything together. So, allow a predicate to be passed to determine whether a file can be deleted. We only let empty __init__.py files be deleted in this case, and by default the original behavior is retained.

Reviewed By: cjhopman

shipit-source-id: f8955806e3