Add KotlinTestCompiler
Summary:
Right now we have a Java TestCompiler class (https://fburl.com/diffusion/wm5ylzw5) that we use in various tests, including StubJarTest. This adds something similar for Kotlin, and will be used in StubJarTest in future diffs.

Right now it is a basic implementation without a lot of methods/parameters that just does compilation of the files that you pass it. We can add more customization in future as and when we need it.

Reviewed By: asp2insp

shipit-source-id: 882b9c1c14