Add basic Kotlin test to StubJarTest
Summary:
StubJarTest contains many tests to verify that we're creating the stubs that we expect. This adds support for testing Kotlin stubs as well, and adds the first test - that of an empty class. As we add the right support for Kotlin ABI generation, we will add more tests to show their correctness.

An argument could be made for having a separate Kotlin test, but given that we're keeping the code for generating ABIs in the same place, I think it's good to also have the tests in the same place (plus it avoids code duplication to a reasonable extent).

Reviewed By: jkeljo

shipit-source-id: ac71904d59