Create XcodeWriteOptions class to encapsulate a bunch of Project metadata
Summary: In order to get a unified output, I need to begin to make the project generator class easier to manipulate. Because of having a ton of parameters, this makes it unruly. As a result, I'm creating a top-level class that will have some high level items that make it easy to pass around and change the inputs of ProjectGenerator.

Reviewed By: williamtwilson

shipit-source-id: 63253611ba