Delete custom target groups of NewNativeTargetProjectMutator
Summary: This var allowed us to specify custom paths we wanted to follow; I think this was used for combined-project, which is a flag we've killed off, making this unused. Given that we don't want to support this functionality any more and to generate the file paths relative to the project cell, we can kill this off.

Reviewed By: williamtwilson

shipit-source-id: 1df7ac408b