Put all files in a project in a directory structure relative to the project cell per project (for now)
Summary:
In the current version of Buck project for Xcode, for each Buck package we create an XcodeProject with the package name and then put all of its files relative to that. For source files, we put them in a local `Sources` directory which does not exist on disk, and then build the local file structure under that.

As part of this overhaul, this diff creates a local folder hierarchy within each individual buck package project (this will go away later once we only have one project per target). This diff handles the sources phase as well as the info plist, prefix headers, and bridging headers. The other ones wil be handled in separate diffs after this one.

I chose the approach here of creating the group path relative to each file instead of in a top level method which creates the structure because I don't really want to make any assumptions on where files are coming from. This ends up being naive, but it at least allows us to be correct more easily. This will also make it easier to handle cases where files are generated in `buck-out` instead of as a subdirectory fo the current Buck package / target; in this case, the `buck-out` directory in the root of a cell will be created properly as each file / subfolder is written to there.

This will also make it easier in the future to kill of the .xcworkspace creating groups; this will allow all group creation to happen in one central location going forward.

Reviewed By: williamtwilson

shipit-source-id: cc0509d459