Move BUCK / build file generation to Xcode Project to NewNativeTargetMutator
Summary: As of now, the BUCK file is added to the xcode project in ProjectGenerator, outside of the class that is doing the rest of the pbxproject generation (NewNativeTargetMutator). This moves this functionality inside of the NewNativeTargetMutator so we are colocating this functionality and maintaining consistency across classes. This also allows us to easily put the Buck file in the proper file directory, reusing all of the previous functionality in the previous diffs.

Reviewed By: williamtwilson

shipit-source-id: 0fafc55ec6