Replaced simctl list with idb list-targets
Summary: We are creating a function that installs a bundle in the simulator using idb instead of simctl. In this diff, the function can find the correct simulator to install the bundle. In order to activate this option, a configuration tag should be added to the buck install command (--config "apple.use_idb=true").

Reviewed By: williamtwilson

shipit-source-id: 81a7d3ce3d