Benchmark DefaultProjectFilesystem.getDirectoryContents
Summary:
I want to optimize DefaultProjectFilesystemView.getDirectoryContents. Write some benchmarks for that function to make development easier and to demonstrate improvements.

Note: This is the first benchmark using JMH. (Existing benchmarks use Caliper, but Caliper is broken for several reasons right now, so I chose the path of least resistance.)

Reviewed By: sbalabanov

shipit-source-id: 60f1193378