Optimize DefaultProjectFilesystemView.getDirectoryContents
Summary:
MorePaths.relativize is relatively CPU-intensive and allocation-intensive. This slows down DefaultProjectFilesystemView.getDirectoryContents. Optimize that function by avoiding MorePaths.relativize. Also switch to Stream, which seems to performs better than FluentIterable for getDirectoryContents.

Also, instead of returning an ImmutableSortedList, return an ImmutableSortedSet. Benchmarking getDirectoryContents showed that constructing a ImmutableSortedSet is faster than constructing a ImmutableSortedList. I did not measure how the performance of consumers are affected by this change in data structure.

In my testing, this makes getDirectoryContents about 55% faster for a directory with 10'000 files.

This diff should not change behavior.

Reviewed By: sbalabanov

shipit-source-id: f4220e6bf9