Add a Xcode script build phase to build with buck
Summary:
This change generates an Xcode project with each target having a script build phase to build with buck added.

In order to achieve this:
1. Add a bash and python script to be called from Xcode, build with buck and copy the generated artifacts to the location Xcode expects them to be
2. Add buck config entries to customize where the scripts are located
3. Add the build phase to the target
4. Add to the target's xcconfig, build settings to export on build

Note, while this adds the build phase to buck, it does not work off the bat. There are two changes that will be performed in followup diffs that need to be made in order for Xcode to build with buck. Namely,
1. Remove unnecessary build phases for targets that generate the same resources, etc.
2. Update schemes such that on run, only the scheme's primary target is built

Additionally, for compilation to work correctly, header maps must be merged in order to reduce the number and size of arguments exported by Xcode during compile. Otherwise the command is too long to execute.

Reviewed By: williamtwilson

shipit-source-id: e1218f4e2e