Update logic on what build actions we perform for each target in a scheme
Summary:
Previously, we were setting build actions for targets on almost all possible "build for" actions. This means that when building with buck, there are multiple targets that attempt to get built at the same time. This is a problem for two reasons:
1. Target A may build Target B as a dependency, meaning there is duplicative work performed
2. Target B may not be compatible with the build script and fail

Instead, updated the build actions for the scheme to be as follows:

For a given target in a scheme
1. always analyze
2. if a test product type (unit or UI), enable for testing
3. if the target is the scheme's primary target,
 a. enable for running
 b. if an application/app-extension/watch-extension product type, enable for profiling and archiving

Reviewed By: williamtwilson

shipit-source-id: b355206c54