Add new markers for description args to support test protocol
Summary:
test rules will get an attribute of specs where the test protocol
information is specified. This is free form for now until we settle on the fields.

They also get an attribute of runner_library where they can specify a library
that contains a runner that will be packed for the test binary

Reviewed By: philipjameson

shipit-source-id: 99295ee427