Don't fail tests if Watchman log file is missing
Summary:
D16876042 caused tests to fail if, for whatever reason, Watchman's log file is inaccessible. For example, if Watchman crashes on start, the test might fail with a FileNotFoundException for Watchman's log file, obscuring the real issue.

When dumping Watchman's logs, report errors read the log file, but don't make the test fail.

Reviewed By: jtorkkola

shipit-source-id: 2ca47027a9