Add KotlinMetadataReader to find inline functions of a file
Summary:
Kotlin methods can be inlined, which means that we need to retain those method bodies as part of the ABI so that the Kotlin compiler can use the method body when it needs to.

Which methods are inlined is stored in the Kotlin metadata annotation, so first we find that annotation and parse it to find the methods whose bodies we need to retain. If there are such methods then we omit the `SKIP_CODE` flag when we visit the class (since that flag will cause us to omit all the method bodies). AbiFilteringClassVisitor is changed to check the list of methods whose bodies we need to retain, and it is also responsible for removing the bodies of all other methods, in the case that we did not use the `SKIP_CODE` flag.

I'm not sure how well I've explained all this so feel free to seek clarification. Hopefully the two test cases in StubJarTest will shed some light.

Reviewed By: jkeljo

shipit-source-id: 4f0315bcbe