Tests for Kotlin compiler generating ABIs
Summary:
The Kotlin compiler has a plugin that generates ABIs: https://youtrack.jetbrains.net/issue/KT-25128. This adds that plugin to KotlinTestCompiler and allows us to run tests in `SOURCE` mode for Kotlin.

The ABIs that the plugin produces are a little different to what we were producing for class ABIs before (they leave in some of the debug code). We want to use what the plugin produces as the source of truth for now, so this also includes a change to the class ABI generation for Kotlin so that it doesn't strip out the debug code for any of the Kotlin classes.

Reviewed By: asp2insp

shipit-source-id: fde35c281d