Prebuilt aar exclude from native lib packaging optimizations
Summary:
When `.aar` files contain both a prebuilt native lib (`.so`) and java code which references this native lib, they usually have an embedded call to `System.loadLibrary()`.
This means that the following optimizations are incompatible with this `.aar`:

* exopackage-native
* packaging libs as `assets/`
* compressing libs via `xz`

This diff adds a flag which can be specified on an `android_prebuilt_aar` to opt out of these optimizations and package native libs directly in the main `.apk`.

Reviewed By: styurin

shipit-source-id: f8f0d51d42