Move File writing logic out of NewNativeTargetProjectMutator and into its own class
Summary:
This logic was originally encapsulated here as a continuation of existing functionality of this class. However, it is becoming apparent that this functionality may be needed in broader constants instead of being confined to `NewNativeTargetMutator` directly (specifically, the next diff). As a result, I'm abstracting this logic into its own class, to reduce the complexity of `NewNativeTargetMutator` as well as increase readability and reusability.

There should be not change to existing functionality.

Reviewed By: williamtwilson

shipit-source-id: afbf2a661d