Move gen rule file generation to NewNativeTargetMutator
Summary:
Files that quality as `GenRule` sources (e.g. adt, remodel, etc.) end up in a created `Others` directory. We should just put them where they are found on disk, which is what this diff does.

Given the complexity in how the gen rule files are handled (they are aggregated through all sub targets and added to the project at the end of generation), I decided to leave the file generation in ProjectGenerator instead of moving it to `NewNativeTargetMutator` like previous diffs did. This is where the use case for the `ProfileFileWriter` class comes in -- we can simply reuse the functionality here easily, with good readability, to get the desired effect.

Reviewed By: williamtwilson

shipit-source-id: f41aed6ef0