Move entitlements file from project root to group representing place on disk
Summary:
As titled. Right now the entitlements file just goes to whichever target group is input (which, in this instance, is the project main group). This uses the ProjectFileWriter class to put it in a group tree similar to the one on disk.

There was actually no unit test for this behavior, so I went ahead and added one as well.

Reviewed By: williamtwilson

shipit-source-id: 467fbaf5cf