Kill off explicit SceneKit setting in ProjectGenerator
Summary: Turns out... we don't need this! Because the SceneKit stuff is already considered an `AppleWrapperResourceArg`, it automatically gets added as a file reference in Xcode as part of the resources phase, which are already being handled as per a previous diff. Xcode seemingly understands how to read a scene kit file extension and just adding the file reference alone is enough to make Xcode do the right thing and show all underlying resouirces that are part of it.

Reviewed By: williamtwilson

shipit-source-id: 40819c0ecf