Stop creating groups in the workspace for the projects
Summary: For each project generated, the workspace code now creates the path to the buck package. So if the cell root is `/MyBuckCell` and the buck file for a package is located in `/MyBuckCell/MyApp/Libs:Lib` then the workspace would build out a folder structure of `/MyApp/Libs/Lib.xcodeproj`. We don't want this behavior any more going forward, so this diff simply removes that behavior. This means that all of the project files end up in the root, but that's ok for the time being as the next diff will migrate us to one simple project.

Reviewed By: williamtwilson

shipit-source-id: 0f1a93e4ce