Manage build target focus outside of the ProjectGenerator
Summary:
Right now, the focus matcher behavior is confusing and not really consistent. This is because it is just passed around ot ProjectGenerator which checks it for somethings and not for other things. This has become problematic in the process of creating one project because we're creating tons of build targets and only adding them in the case of the focus flag in the ProjectGenerator, which is just way too much extra work -- we don't need to  build tons of unfocused build targets anymore because buck is going to build behind the scenes.

This cleans up the focused logic by throwing it in the top level of the WorkspaceAndProjectGenerator and no longer passing it to the project generator. We can simply do a lot less work filtering the build targets at the top level and then only handling the things we want.

Reviewed By: williamtwilson

shipit-source-id: 9eae10b2de