Build project data relative to the project cell instead of the target cell
Summary: In testing, I've noticed projects outside of the current project cell not rendering properly. This is because the relative paths are not resolving properly as a result of using a shared project path that exists in one cell, yet the project generator loop using relative paths to the cell the project falls in. Given how we're unifying all of this stuff, this doesn't really need to be done anymore, and it cleans up a lot of the code in this method since we no longer have to group stuff as much -- we can simply associate paths to build targets and go from there.

Reviewed By: williamtwilson

shipit-source-id: 146583a7d6