Cast a wider net when searching for code coverage files
Summary:
Right now test coverage for java will only look at the classpath of the test. This excludes a common
pattern where sources are compiled, post-processed, then the resulting jar is placed on the classpath:

```
java_library(
  name = "foo",
  srcs = [ "Foo.java"],
  tests = [ ":foo_test" ],
)

genrule(
  name = "foo_processed_gen",
  cmd = "do_something $(location :foo) > $OUT",
  out = "processed.jar",
)

prebuilt_jar(
  name = "foo_processed",
  binary_jar = ":foo_processed_gen",
)

java_library(
  name = "top_level",
  deps = [ ":foo_processed"],
)

java_test(
  name = "wants_to_test_foo",
  deps = [":top_level"],
)
```

In this instance, it's not correct for `top_level` or `wants_to_test_foo` to depend on `foo` because we actually want the post-processed classes. But we still need to be able to find `Foo.java` when the time comes to gather a sourcepath for coverage calculation.

This diff expands the search to include all transitive build deps of the test, even if they aren't directly on the classpath so we can find source files like the above example.

Reviewed By: jtorkkola

shipit-source-id: e899559d17