rust: implement mapped sources
Summary:
This adds `mapped_srcs = { "path": "mapped_path" }` to Rust rules.
This allows the local name for a source file to be defined independent
of the actual source path. This is mostly useful when a source file
is generated by a rule, as it will be generated in a completely
different rule-specific directory. With this feature it can be
mapped to the appropriate place where module structure of the crate
expects it.

Reviewed By: cjhopman

shipit-source-id: 58595a0a5d