rust: add named_deps attribute for rules
Summary:
This allows the local name of a crate dependency to be set via
`named_deps = { "localname": ":some-crate" }`. This is primarily
useful in allowing a crate to depend on multiple crates with the
same name (such as using multiple versions of the same crate, or
name collisions caused by naming conventions).

Reviewed By: cjhopman

shipit-source-id: 7483788de6