Multitenant: attach universe to getPotentiallyAffectedBuildPackages
Summary:
To determine which package was modified based on a list of changed files, we need to know the locations of all the packages in projected universe (= corpus). This change exposes a function in an index which reports whether a package with specific path exists in the universe or not.

Finally, this function is hooked to a component that detects changed packages to rebuild.

Not that with resilient parsing the universe is now properly populated with the packages even if some packages cannot be parsed.

Reviewed By: mykola-semko

shipit-source-id: 7442e69d74