Refactor FsToBuildPackageChangeTranslator to decouple it from the service
Summary:
It seems that the design, where FsToBuildPackageChangeTranslator is a long-living object being instantiated with the service does not really work. The reason is that in order to translate changes it needs to know some basic information about current state, such as the universe of all the packages and build rules for the base generation.

This change factors out FsToBuildPackageChangeTranslator from the index and allows consumers to instantiate the specific implementation at-place. There supposed to be 2 implementations:

1) The one used with state creator - it will use an index of already parsed packages as the universe and calls buck command line to parse changed packages

2) The one used with SCMTailer - it will use GenerationMap as an index of based packages and uses SCMQueryFilesystem to parse packages.

Reviewed By: mykola-semko

shipit-source-id: 3c29cfdd48