Include toolchain/usr/lib/swift in linker paths
Summary:
In D16585218 the runtime paths for swift were changed to include only
those that contained the `libSwiftCore.dylib` but as it turns out the
linker also needs a path to the new `libswiftCompatibility50.a` and
`libswiftCompatibilityDynamicReplacements.a` modules also. In the
current Xcode 11 betas those appear in the toolchain/usr/lib/swift
directory.

Reviewed By: b3ll

shipit-source-id: 8ce54913c7