Avoid storing intermediate files in a temp directory when building zip_file
Summary:
This change introduces an alternative implementation of `zip_file` rule that avoids keeping intermediate data on disk and provides more flexible control over duplicate entries.

On user level this change introduces a new option for `zip_file`: `on_duplicate_entry`. This option defines what should happen when a zip file is created with multiple entries with the same name. Possible values are "fail", "overwrite" and "append".

The implementation is organized in two phases: collecting information about source entries and writing data to the output zip file using the information collected in a previous step.

`ZipEntrySourceCollection` represents the collection of sources for a zip file. A source for an entry can be either a file or an entry in another zip file. `ZipEntrySourceCollectionBuilder` is responsible for handling input files and archives, dealing with duplicate entries and keeping entries in the right order.

When the collection is created it is written to the output file by `ZipEntrySourceCollectionWriter`.

Reviewed By: bobyangyf

shipit-source-id: a54c08cfb1