Use entitlements during code signing stage on macOS
Summary:
Before, for platforms on which `adHocCodeSignIsSufficient()`:

```
 name.startsWith(IPHONESIMULATOR.getName())
        || name.startsWith(WATCHSIMULATOR.getName())
        || name.startsWith(APPLETVSIMULATOR.getName())
        || name.startsWith(MACOSX.getName());
```

entitlements were just skipped during code signing phase. That included macOS where entitlements should be a part of notarization process. So on macOS entitlements were just not taken into account at all.

In this diff I am changing the behavior to check regardless of the platform if the entitlements are provided and use them later in the code signing stage.

That's an opt-in behavior which is turned by `--config apple.use_entitlements_when_adhoc_code_signing=true`.

Reviewed By: milend

shipit-source-id: c321beef54