RawTargetNodePipeline pipeline should use unconfigured node as key
Summary:
**Summary**

Essentially, this diff changes return type of of `RawTargetNodePipeline.getBuildTarget` from `BuildTarget` to `UnconfiguredBuildTargetView` and removes **TODO** from that function.

**Why**

This is important to better understand what's going on in the parser pipeline: `RawTargetNodePipeline` produces unconfigured targets, not configured targets.

**Future considerations**

Ideally code should work with `UnconfiguredBuildTarget`, not with `UnconfiguredBuildTargetView`, this is something I should clean up later.

Reviewed By: styurin

shipit-source-id: 3d9caa3115