Parsing and presenting the idb result
Summary: In the previous diff we got buck to execute the test using idb. Now we got it to understand the result and log it. The only thing missing is the printing of the final result. Something we also have to work on is to make idb print when it is starting the tests, since buck needs that to log it. It is currently not being called due to the issues we have to fix. The implementation used was similar to the xctool way.

Reviewed By: zeyadsalloum

shipit-source-id: 69c728cda4