Allow Artifact to be stringified into an absolute path
Summary: Sometimes we need to have the absolute path to an artifact e.g. runtime.exec needs it because it does not change to `directory(Path)` until /after/ it has resolved the exectuable path. However, we don't /always/ want it. Additional arguments beyond the first one (if its an artifact) passed to users should always use paths relative to the working dir.

Reviewed By: bobyangyf

shipit-source-id: 2dc887465e