Materialize non-classpath rules-under-test for JUnit tests
Summary:
Now that we consider some rules which are not on the direct classpath to be candidates for rules-under-tests we need to ensure those rules are materialized correctly
so they can be analyzed. This diff adds the rules-under-tests to the build directly. We also fix a missing call to `filesystem.getPathForRelativePath` that would cause extracting a jar
file fetched from cache to fail.

Reviewed By: IanChilds

shipit-source-id: 9b7abf8cfb