Generate only one project for all targets
Summary:
With this diff, we finally consolidate all of the previous work to generate just one xcode project with a file hierarchy similar to the one found on disk. To accomplish this, I streamlined a bunch of APIs:

- Killed off the `ProjectGeneratorMap` with the path to project generator mapping. It was a poor choice for a lot of reasons -- it  is a mutable parameter passed all around that gets mutated -- really just not a great pattern; also it exposes implementation details, which we don't want
- Because of the above, I created `Result` objects that are returned with details of things we want to inspect; this is a pattern I'd like to move towards in the future, where we are returning information about the output, rather than specific things we care about; for example, we returned a `Path` from the workspace generator method which pointed to the workspace path, but this isn't terribly clear (it could be a path to anything!)
- Killed off the futures generation in `ProjectGenerator`; for one, we only need to call this once (for now, more on this later). Second, I don't think it actually parallelized anything because the `generateProjectForDirectory` method which actually ran this async was `synchronized` which basically means it was running serially anyway. oof.

For performance reasons, I think we can do further optimizations here, but this works pretty well as is. The two biggest bottle necks are the target generation (because it's serial and slow) and writing the project file to disk. The next thing I'm going to look into is seeing if we can parallelize this or optimize it in any way to improve buck project times and make it mostly dependent on the buck action graph generation.

Reviewed By: williamtwilson

shipit-source-id: 5973df1647