Break up NewNativeTargetMutator into a builder and writer class
Summary:
This class is very large and is weirdly a kind-of builder but also a generator -- too much going on. This simplifies the logic considerably -- all of the `builder` components have been moved to an Immutable object using Immutable and allows builders. From there, the mutator's only responsibility is taking this as input and writing it to a project file. This cleans up this class considerably, makes its components easier to understand, and fixes some underlying mutability. I also changed the naming to be more reflective of the new responsibilities:

- NewNativeTargetMutator -> `XcodeNativeTargetProjectWriter`
- (builder components of NewNativeTargetMutator) -> `XCodeNativeTargetDescription`

Looking forward, i can see the `XCodeNativeTargetDescription` being a return value from ProjectGenerator, which would allow us to parallelize a lot of the build target generation safely as the underlying mutations would be removed finally.

Reviewed By: williamtwilson

shipit-source-id: e51b5fbbe0