Move Target Dependency setting outside of createBinaryTarget method in ProjectGenerator
Summary:
This is the beginning of some cleanup I'm looking to do to simplify and streamline a lot of the logic within `ProjectGenerator`. The biggest thing is I would like to parallelize some of the work going on in ProjectGenerator which is right now not possible given how mutable and stateful everything is.

One thing that is tricky is handling deps between targets -- for example, test targets needing the main target as a dep in order to function properly (maybe? we may not need this anymore depending on future work). Anyway, I think we should move this out of ProjectGenerator and handle this above because it makes no sense that this class can generate binary targets but then also be able to explicitly know about every _other_ binary target ever built and then associate deps with them.

Step one here is simply extracting this logic outside of the `generateBinaryTarget` function. Once I'm done, I envision this code will simply return stuff it needs to depend on, but the actual dependency handling will be handled outside of this class.

Reviewed By: williamtwilson

shipit-source-id: 75157d1e48