Move genrule file writing out of ProjectGenerator
Summary: Genrule files were a bit tricky before, which is why I punted on them. The issue was that they were not added as part of a target, just aggregated from the gen rule target and then added after all targets were compiled into a binary. Now that we're breaking this down into handling each target individually, this logic breaks a bit. This diff adds the gen rule files as expected to the NativeTargetAttributes object, and modifies some behavior, namely that writing a target doesn't necessarily involve adding a PBXNativeTarget as well, which required some modifications of behavior.

Reviewed By: williamtwilson

shipit-source-id: eb841483ed