Remove required build targets side effect from resolveSourcePath function in ProjectGenerator
Summary:
This is a really nasty side effect in a method that should only ever eb returning paths -- not adding to an array that is checked later. This blocks us from making the core of ProjectGenerator immutable and stateless as we need to make `requiredBuildTargetsBuilder` a local variable instead of a class field.

To resolve this, this diff centralizes the logic to write required build targets in a method that can check all SourcePaths, as well as a few other spurious places in ProjectGenerator; this allows us to completely remove this logic out of `resolveSourcePath` which should reduce the statefulness considerably.

Reviewed By: williamtwilson

shipit-source-id: ded4a7e075