Added unconfigured selector list resolver
Summary:
Added a new selector list resolver which acts the same as the default one except for the times when the values of the select dictionary are lists, in which case it concatenates all of them in order to have information on all configurations.

`AbstractParsingContext.java` is a generator for `ParsingContext.java`.

When a `ParsingContext` object is created, the instance will now have a new callable method, `withUseUnconfiguredSelectorResolver`, which is used to set the internal field `useUnconfiguredSelectorResolver`. The method with the same name as this field is called in order to check is value it is used in order to decide between the two selector list resolvers available: `DefaultSelectorListResolve` and `UnconfiguredSelectorListResolver`.

Reviewed By: stepancheg

shipit-source-id: e63613867f