Ensure min-api-version is passed to dx when we're aware of it
Summary:
dx will produce different output when targeting a device which supports it.
In order to take advantage of this, we need to ensure dx knows what the min api level of the app is.
This can be provided in the app via the `manifest_entries` already, so this diff just pipes it
through to the app.

Note that we do not support pre-dexing in this mode, since the at the individual pre-dex library level
we do not know which API version will be used for the final app, so this change restricts the new flag
usage to non-pre-dexed-buildables only.

Reviewed By: IanChilds

shipit-source-id: b220a3fd90