Fix the output of sh_binary() to work when invoked through a symlink
Summary:
The code in `sh_binary_template` attempts to find the buck project root
by looking at the directory that contains `${BASH_SOURCE[0]}`.  This
doesn't work if `${BASH_SOURCE[0]}` is a symlink that points to the real
`sh_binary()` script that is in a different directory.

This fixes the code to properly resolve symlinks before looking at the
parent directory.  Doing so is somewhat difficult to do portably.  For
now I have copied the code currently used by the `make_par` utility for
running Python scripts.

Reviewed By: philipjameson

shipit-source-id: 8c7f70f9d2